+++ sobolevsky.misha/S5/errorsANDinput.hpp
@@ -5,7 +5,7 @@
 
 namespace sobolevsky
 {
-  void inputFromFile(std::istream &in, sobolevsky::AVLtree< int, std::string, int > &data);
+  void inputFromFile(std::istream &in, sobolevsky::AVLtree< long long, std::string, long long > &data);
   void errorInvalidCommand(std::ostream &out);
   void errorEmpty(std::ostream &out);
   void errorOverflow(std::ostream &out);
